♻️ App/server terminology, refactor structure
