+++ src/index.ts
@@ -1,4 +1,4 @@
-export { Server } from './lib/server'
+export { ProtoCat } from './lib/server/application'
 export { CallType } from './lib/call-types'
-export { onError } from './lib/middleware/on-error'
-export { Middleware } from './lib/call'
+export { onError } from './lib/server/middleware/on-error'
+export { Middleware, ServiceImplementation } from './lib/server/call'
