+++ src/lib/server/middleware/compose-middleware.ts
@@ -1,4 +1,4 @@
-import { Middleware } from './call'
+import { Middleware } from '../call'
 
 export const composeMiddleware = (middleware: Middleware[]): Middleware => {
   return async (call, next) => {
